+++ src/__tests__/modules/quickpick.test.ts
@@ -10,17 +10,18 @@ export type Item = QuickPickItem | string
 
 let nvim: Neovim
 let disposables: Disposable[] = []
+let ns: number
 
 beforeAll(async () => {
   await helper.setup()
   nvim = helper.nvim
+  ns = await nvim.createNamespace('coc-input-box')
 })
 afterAll(async () => {
   await helper.shutdown()
 })
 
 afterEach(async () => {
-  await helper.reset()
   disposeAll(disposables)
   disposables = []
 })
@@ -34,6 +35,109 @@ async function getTitleLine(): Promise<string> {
   let lines = await buf.lines
   return lines[0]
 }
+
+describe('InputBox', () => {
+  it('should request input', async () => {
+    let winid = await nvim.call('win_getid')
+    let p = window.requestInput('Name')
+    await helper.waitFloat()
+    await nvim.input('bar<enter>')
+    let res = await p
+    let curr = await nvim.call('win_getid')
+    expect(curr).toBe(winid)
+    expect(res).toBe('bar')
+  })
+
+  it('should use input method of vim', async () => {
+    helper.updateConfiguration('coc.preferences.promptInput', false)
+    let defaultValue = 'default'
+    let p = window.requestInput('Name', defaultValue)
+    await helper.wait(50)
+    await nvim.input('<enter>')
+    let res = await p
+    expect(res).toBe(defaultValue)
+  })
+
+  it('should return empty string when input empty', async () => {
+    let p = window.requestInput('Name')
+    await helper.wait(30)
+    await nvim.input('<enter>')
+    let res = await p
+    expect(res).toBe('')
+  })
+
+  it('should emit change event', async () => {
+    let input = await window.createInputBox('', '', {})
+    disposables.push(input)
+    let curr: string
+    input.onDidChange(text => {
+      curr = text
+    })
+    await nvim.input('abc')
+    await helper.waitValue((() => {
+      return curr
+    }), 'abc')
+    input.title = 'foo'
+    expect(input.title).toBe('foo')
+    input.loading = true
+    expect(input.loading).toBe(true)
+    input.borderhighlight = 'WarningMsg'
+    expect(input.borderhighlight).toBe('WarningMsg')
+  })
+
+  it('should not check bufnr for events', async () => {
+    let input = await window.createInputBox('', undefined, {})
+    disposables.push(input)
+    let bufnr = input.bufnr
+    let called = false
+    input.onDidChange(() => {
+      called = true
+    })
+    await events.fire('BufWinLeave', [bufnr + 1])
+    await events.fire('PromptInsert', ['', bufnr + 1])
+    await events.fire('TextChangedI', [bufnr + 1, {
+      lnum: 1,
+      col: 1,
+      line: '',
+      changedtick: 0,
+      pre: ''
+    }])
+    expect(called).toBe(false)
+    expect(input.bufnr).toBeDefined()
+    expect(input.dimension).toBeDefined()
+  })
+
+  it('should change input value', async () => {
+    let input = await window.createInputBox('', undefined, {})
+    disposables.push(input)
+    let called = false
+    input.onDidChange(() => {
+      called = true
+    })
+    input.value = 'foo'
+    await helper.waitValue(async () => {
+      let lines = await nvim.call('getbufline', [input.bufnr, 1]) as string[]
+      return lines[0]
+    }, 'foo')
+    expect(called).toBe(true)
+    expect(input.value).toBe('foo')
+  })
+
+  it('should show and hide placeHolder', async () => {
+    let input = await window.createInputBox('title', undefined, { placeHolder: 'placeHolder' })
+    disposables.push(input)
+    let buf = nvim.createBuffer(input.bufnr)
+    let markers = await buf.getExtMarks(ns, 0, -1, { details: true })
+    expect(markers.length).toBe(1)
+    let blocks = markers[0][3].virt_text
+    expect(blocks).toEqual([['placeHolder', 'CocInputBoxVirtualText']])
+    await nvim.input('a')
+    await helper.waitValue(async () => {
+      let markers = await buf.getExtMarks(ns, 0, -1, { details: true })
+      return markers.length
+    }, 0)
+  })
+})
 
 describe('QuickPick', () => {
   it('should not thrown when window not shown', async () => {
@@ -156,6 +260,68 @@ describe('showQuickPick', () => {
   })
 })
 
+describe('QuickPick configuration', () => {
+  afterEach(() => {
+    helper.workspace.configurations.reset()
+  })
+
+  it('should respect width of quickpick', async () => {
+    helper.updateConfiguration('dialog.maxWidth', null)
+    let quickpick = await window.createQuickPick()
+    disposables.push(quickpick)
+    quickpick.items = [{ label: 'foo' }, { label: 'bar' }]
+    quickpick.width = 50
+    quickpick.value = ''
+    await quickpick.show()
+    let win = nvim.createWindow(quickpick.winid)
+    let width = await win.width
+    expect(width).toBe(50)
+  })
+
+  it('should scroll by <C-f> and <C-b>', async () => {
+    helper.updateConfiguration('dialog.maxHeight', 2)
+    let quickpick = await window.createQuickPick()
+    quickpick.value = ''
+    quickpick.items = [{ label: 'one' }, { label: 'two' }, { label: 'three' }]
+    disposables.push(quickpick)
+    await quickpick.show()
+    let winid = quickpick.winid
+    await nvim.input('<C-f>')
+    await helper.wait(1)
+    await nvim.input('<C-f>')
+    await helper.waitValue(async () => {
+      let info = await nvim.call('getwininfo', [winid])
+      return info[0].topline
+    }, 2)
+    await nvim.input('<C-b>')
+    await nvim.input('<C-x>')
+    await helper.wait(1)
+    await nvim.input('<C-b>')
+    await helper.waitValue(async () => {
+      let info = await nvim.call('getwininfo', [winid])
+      return info[0].topline
+    }, 1)
+  })
+
+  it('should respect configurations', async () => {
+    helper.updateConfiguration('dialog.maxWidth', 30)
+    helper.updateConfiguration('dialog.rounded', false)
+    helper.updateConfiguration('dialog.floatHighlight', 'Normal')
+    helper.updateConfiguration('dialog.floatBorderHighlight', 'Normal')
+    helper.updateConfiguration('dialog.maxHeight', 2)
+    let quickpick = await window.createQuickPick()
+    quickpick.items = [{ label: 'one' }, { label: 'two' }, { label: 'three' }]
+    await quickpick.show()
+    let winids = await nvim.call('coc#float#get_float_win_list') as number[]
+    let winid = Math.max(...winids)
+    let win = nvim.createWindow(winid)
+    let h = await win.height
+    expect(h).toBe(2)
+    await nvim.input('<esc>')
+  })
+
+})
+
 describe('createQuickPick', () => {
   it('should throw when unable to open input window', async () => {
     let fn = nvim.call
@@ -177,20 +343,20 @@ describe('createQuickPick', () => {
 
   it('should throw when unable to open list window', async () => {
     let fn = nvim.call
-    nvim.call = (...args: any) => {
+    let spy = jest.spyOn(nvim, 'call').mockImplementation((...args: any) => {
       if (args[0] === 'coc#dialog#create_list') return undefined
       return fn.apply(nvim, args)
-    }
-    disposables.push(Disposable.create(() => {
-      nvim.call = fn
-    }))
+    })
     let fun = async () => {
       let quickpick = await window.createQuickPick({
         items: [{ label: 'foo' }, { label: 'bar' }],
       })
+      disposables.push(quickpick)
       await quickpick.show()
     }
     await expect(fun()).rejects.toThrow(/Unable to open/)
+    spy.mockRestore()
+    await nvim.call('feedkeys', [String.fromCharCode(27), 'in'])
   })
 
   it('should respect initial value', async () => {
@@ -204,43 +370,6 @@ describe('createQuickPick', () => {
     let lines = await buf.lines
     expect(lines[0]).toBe('value')
     await nvim.input('<esc>')
-  })
-
-  it('should respect width of quickpick', async () => {
-    helper.updateConfiguration('dialog.maxWidth', null)
-    let q = await window.createQuickPick()
-    q.items = [{ label: 'foo' }, { label: 'bar' }]
-    q.width = 50
-    q.value = ''
-    await q.show()
-    let win = nvim.createWindow(q.winid)
-    let width = await win.width
-    expect(width).toBe(50)
-  })
-
-  it('should scroll by <C-f> and <C-b>', async () => {
-    helper.updateConfiguration('dialog.maxHeight', 2)
-    let quickpick = await window.createQuickPick()
-    quickpick.value = ''
-    quickpick.items = [{ label: 'one' }, { label: 'two' }, { label: 'three' }]
-    await quickpick.show()
-    disposables.push(quickpick)
-    let winid = quickpick.winid
-    await nvim.input('<C-f>')
-    await helper.wait(1)
-    await nvim.input('<C-f>')
-    await helper.waitValue(async () => {
-      let info = await nvim.call('getwininfo', [winid])
-      return info[0].topline
-    }, 2)
-    await nvim.input('<C-b>')
-    await nvim.input('<C-x>')
-    await helper.wait(1)
-    await nvim.input('<C-b>')
-    await helper.waitValue(async () => {
-      let info = await nvim.call('getwininfo', [winid])
-      return info[0].topline
-    }, 1)
   })
 
   it('should change current line by <C-j> and <C-k>', async () => {
@@ -265,6 +394,7 @@ describe('createQuickPick', () => {
   it('should toggle selected item by <C-space>', async () => {
     let quickpick = await window.createQuickPick()
     quickpick.items = [{ label: 'one' }, { label: 'two' }, { label: 'three' }]
+    await quickpick.show()
     disposables.push(quickpick)
     await nvim.input('<C-sapce>')
     await helper.wait(10)
@@ -285,23 +415,6 @@ describe('createQuickPick', () => {
     await events.fire('BufWinLeave', [quickpick.buffer.id + 1])
     await events.fire('PromptKeyPress', [quickpick.buffer.id + 1, 'C-f'])
     expect(quickpick.currIndex).toBe(0)
-  })
-
-  it('should respect configurations', async () => {
-    helper.updateConfiguration('dialog.maxWidth', 30)
-    helper.updateConfiguration('dialog.rounded', false)
-    helper.updateConfiguration('dialog.floatHighlight', 'Normal')
-    helper.updateConfiguration('dialog.floatBorderHighlight', 'Normal')
-    helper.updateConfiguration('dialog.maxHeight', 2)
-    let quickpick = await window.createQuickPick()
-    quickpick.items = [{ label: 'one' }, { label: 'two' }, { label: 'three' }]
-    await quickpick.show()
-    let winids = await nvim.call('coc#float#get_float_win_list') as number[]
-    let winid = Math.max(...winids)
-    let win = nvim.createWindow(winid)
-    let h = await win.height
-    expect(h).toBe(2)
-    await nvim.input('<esc>')
   })
 
   it('should change title', async () => {